requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
dateutils==0.6.12
Django==5.0.6
django-cors-headers==4.4.0
django-filter==24.2
django-redis==5.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
Events==0.5
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Levenshtein==0.25.1
Markdown==3.6
packaging==24.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.3
redis==5.0.6
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
six==1.16.0
sqlparse==0.5.0
swagger-spec-validator==3.0.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
gunicorn==23.0.0
urllib3==1.26.19
waitress
